requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
langchain>=0.0.267
langchain-google-genai>=0.0.5
langserve[all]==0.0.30
pydantic>=1.10.8,<2.0.0
sse_starlette
python-dotenv>=1.0.1
httpx<0.26  # Prevent ImportError from langserve
